+++ server/src/controllers/pessoaController.js
@@ -1,19 +1,20 @@
 const Pessoa = require("../models/pessoas");
 const Cidade = require("../models/cidades");
 const Tipo = require("../models/tipoSanguinio");
+const Doacoes = require("../models/doacoes");
+
 module.exports = {
   async create(request, response) {
     const { id, nome, rua, numero, complemento, documento, cidadeId, tipoId } =
       request.body;
 
-    if (id === "" || id == null) {
+    if (id === "" || id === null) {
       return response.status(401).json({ error: "Id Vazio" });
     }
     //Verifica se existe a cidade, se existe o tipo, e se a pessoa ainda nao existe.
-    if (
-      (await Cidade.findOne({ id: cidadeId })) &&
-      (await Tipo.findOne({ id: tipoId }))
-    ) {
+    const tipo = await Tipo.findOne({ id: tipoId });
+    const cidade = await Cidade.findOne({ id: cidadeId });
+    if (tipo && cidade) {
       const pessoa = await Pessoa.findOne({ id: id });
       if (!(await Pessoa.findOne({ id: id }))) {
         const pessoaCreated = await Pessoa.create({
@@ -23,8 +24,8 @@ module.exports = {
           numero,
           complemento,
           documento,
-          cidadeId,
-          tipoId,
+          cidadeId: cidade.id,
+          tipoId: tipo.id,
           created_at: new Date(),
           updated_at: new Date(),
         });
@@ -85,6 +86,11 @@ module.exports = {
 
   async delete(request, response) {
     const { id } = request.params;
+    if (await Doacoes.findOne({ pessoaId: id })) {
+      return response.status(404).json({
+        error: "Pessoa não pode ser deletada pois possui dependencias",
+      });
+    }
     const pessoaDeleted = await Pessoa.findOneAndDelete({ id: id });
     if (pessoaDeleted) {
       return response.json(pessoaDeleted);
